refs #26 IDの型をLongに統一する

+++ sample-domain/src/main/java/com/sample/domain/dao/system/PermissionDao.java
@@ -32,7 +32,7 @@ public interface PermissionDao {
      * @return
      */
     @Select
-    Optional<Permission> selectById(Integer id);
+    Optional<Permission> selectById(Long id);
 
     /**
      * 権限を1件取得します。
